Refactored feedback components
